updates for comments
