+++ tests/test_config/test_matcha_config.py
@@ -7,82 +7,11 @@ from typing import Dict, Iterator
 import pytest
 
 from matcha_ml.config.matcha_config import (
+    DEFAULT_CONFIG_NAME,
     MatchaConfig,
-    MatchaConfigComponent,
-    MatchaConfigComponentProperty,
     MatchaConfigService,
 )
 from matcha_ml.errors import MatchaError
-
-
-@pytest.fixture
-def mocked_matcha_config_json_object() -> Dict[str, Dict[str, str]]:
-    """A fixture returning a dictionary representation of the matcha.config.json file.
-
-    Returns:
-        matcha_config (Dict[str, Dict[str, str]]): a dictionary representation of the matcha.config.json file
-    """
-    matcha_config = {
-        "remote_state_bucket": {
-            "account_name": "test_account_name",
-            "container_name": "test_container_name",
-            "resource_group_name": "test_resource_group_name",
-        }
-    }
-
-    return matcha_config
-
-
-@pytest.fixture
-def mocked_matcha_config_component_property() -> MatchaConfigComponentProperty:
-    """A fixture returning a mocked MatchaConfigComponentProperty instance.
-
-    Returns:
-        (MatchaConfigComponentProperty): a mocked MatchaConfigComponentProperty instance.
-    """
-    return MatchaConfigComponentProperty(name="account_name", value="test_account_name")
-
-
-@pytest.fixture
-def mocked_matcha_config_component(
-    mocked_matcha_config_json_object: Dict[str, Dict[str, str]]
-):
-    """A fixture returning a mocked MatchaConfigComponentProperty instance.
-
-    Args:
-        mocked_matcha_config_json_object (Dict[str, Dict[str, str]]): a dictionary representation of the matcha.config.json file
-
-    Returns:
-        (MatchaConfigComponentProperty): a mocked MatchaConfigComponentProperty instance.
-    """
-    properties = []
-    for key, value in mocked_matcha_config_json_object["remote_state_bucket"].items():
-        properties.append(MatchaConfigComponentProperty(name=key, value=value))
-
-    return MatchaConfigComponent(name="remote_state_bucket", properties=properties)
-
-
-@pytest.fixture
-def mocked_matcha_config(mocked_matcha_config_component: MatchaConfigComponent):
-    """A fixture returning a mocked MatchaConfig instance.
-
-    Args:
-        mocked_matcha_config_component (MatchaConfigComponent): a mocked MatchaConfigComponent instance.
-
-    Returns:
-        (MatchaConfig): a mocked MatchaConfig instance.
-    """
-    return MatchaConfig([mocked_matcha_config_component])
-
-
-@pytest.fixture
-def mocked_matcha_config_service():
-    """A fixture representing a mocked MatchaConfigService.
-
-    Returns:
-    (MatchaConfigService): a mocked MatchaConfigService instance.
-    """
-    return MatchaConfigService()
 
 
 def test_matcha_config_to_dict(
@@ -133,7 +62,7 @@ def test_matcha_config_service_write_matcha_config(
         mocked_matcha_config (MatchaConfig): a mocked MatchaConfig instance
     """
     matcha_config_file_path = os.path.join(
-        matcha_testing_directory, "matcha.config.json"
+        matcha_testing_directory, DEFAULT_CONFIG_NAME
     )
 
     assert not os.path.isfile(matcha_config_file_path)
@@ -162,7 +91,7 @@ def test_matcha_config_service_read_matcha_config(
         mocked_matcha_config (MatchaConfig): a mocked MatchaConfig instance
     """
     matcha_config_file_path = os.path.join(
-        matcha_testing_directory, "matcha.config.json"
+        matcha_testing_directory, DEFAULT_CONFIG_NAME
     )
 
     with open(matcha_config_file_path, "w") as file:
@@ -204,7 +133,7 @@ def test_matcha_config_service_delete_matcha_config(
         mocked_matcha_config_json_object (Dict[str, Dict[str, str]]): a dictionary representation of the matcha.config.json file
     """
     matcha_config_file_path = os.path.join(
-        matcha_testing_directory, "matcha.config.json"
+        matcha_testing_directory, DEFAULT_CONFIG_NAME
     )
 
     with open(matcha_config_file_path, "w") as file:
@@ -232,7 +161,7 @@ def test_matcha_config_service_delete_matcha_config_error_handling(
         mocked_matcha_config_service (MatchaConfigService): a mocked MatchaConfigService instance
     """
     matcha_config_file_path = os.path.join(
-        matcha_testing_directory, "matcha.config.json"
+        matcha_testing_directory, DEFAULT_CONFIG_NAME
     )
 
     assert not os.path.isfile(matcha_config_file_path)
