+++ tests/test_state/test_remote_state_manager.py
@@ -13,8 +13,6 @@ from azure.mgmt.confluent.models._confluent_management_client_enums import (  # 
 from matcha_ml.config import (
     DEFAULT_CONFIG_NAME,
     MatchaConfig,
-    MatchaConfigComponent,
-    MatchaConfigComponentProperty,
     MatchaConfigService,
 )
 from matcha_ml.errors import MatchaError
@@ -30,64 +28,6 @@ BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 TEMPLATE_DIR = os.path.join(
     BASE_DIR, os.pardir, os.pardir, "src", "matcha_ml", "infrastructure"
 )
-
-
-@pytest.fixture
-def mocked_matcha_config_json_object() -> Dict[str, Dict[str, str]]:
-    """A fixture returning a dictionary representation of the matcha.config.json file.
-
-    Returns:
-        matcha_config (Dict[str, Dict[str, str]]): a dictionary representation of the matcha.config.json file
-    """
-    matcha_config = {
-        "remote_state_bucket": {
-            "account_name": "test-account",
-            "container_name": "test-container",
-            "resource_group_name": "test-rg",
-        }
-    }
-
-    return matcha_config
-
-
-@pytest.fixture
-def mocked_matcha_config_component_property() -> MatchaConfigComponentProperty:
-    """A fixture returning a mocked MatchaConfigComponentProperty instance.
-
-    Returns:
-        (MatchaConfigComponentProperty): a mocked MatchaConfigComponentProperty instance.
-    """
-    return MatchaConfigComponentProperty(name="account_name", value="test_account_name")
-
-
-@pytest.fixture
-def mocked_matcha_config_component(mocked_matcha_config_json_object):
-    """A fixture returning a mocked MatchaConfigComponentProperty instance.
-
-    Args:
-        mocked_matcha_config_json_object (Dict[str, Dict[str, str]]): a dictionary representation of the matcha.config.json file
-
-    Returns:
-        (MatchaConfigComponentProperty): a mocked MatchaConfigComponentProperty instance.
-    """
-    properties = []
-    for key, value in mocked_matcha_config_json_object["remote_state_bucket"].items():
-        properties.append(MatchaConfigComponentProperty(name=key, value=value))
-
-    return MatchaConfigComponent(name="remote_state_bucket", properties=properties)
-
-
-@pytest.fixture
-def mocked_matcha_config(mocked_matcha_config_component):
-    """A fixture returning a mocked MatchaConfig instance.
-
-    Args:
-        mocked_matcha_config_component (MatchaConfigComponent): a mocked MatchaConfigComponent instance.
-
-    Returns:
-        (MatchaConfig): a mocked MatchaConfig instance.
-    """
-    return MatchaConfig([mocked_matcha_config_component])
 
 
 @pytest.fixture
